Refactoring Utils classes
