+++ src/main/java/org/snipsniper/utils/FileUtils.java
@@ -3,6 +3,9 @@ package org.snipsniper.utils;
 import org.snipsniper.LogManager;
 
 import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.PrintWriter;
+import java.util.ArrayList;
 
 public class FileUtils {
     public static boolean delete(String file) {
@@ -60,5 +63,43 @@ public class FileUtils {
             if(!mkdirs(folder))
                 success = false;
         return success;
+    }
+
+    public static File[] listFiles(String folder) {
+        return listFiles(new File(folder));
+    }
+
+    public static File[] listFiles(File folder) {
+        return folder.listFiles();
+    }
+
+    public static ArrayList<String> getFilesInFolders(String path) {
+        ArrayList<String> result = new ArrayList<>();
+        for(File file : FileUtils.listFiles(path)) {
+            if(file.isDirectory())
+                result.addAll(getFilesInFolders(file.getAbsolutePath()));
+            if(!file.isDirectory())
+                result.add(StringUtils.correctSlashes(file.getAbsolutePath()));
+        }
+        return result;
+    }
+
+    public static void printFile(String filename, String text) {
+        try {
+            PrintWriter out = new PrintWriter(filename);
+            out.print(text);
+            out.close();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public static String getFileExtension(File file) {
+        String name = file.getName();
+        int lastIndexOf = name.lastIndexOf(".");
+        if (lastIndexOf == -1) {
+            return ""; // empty extension
+        }
+        return name.substring(lastIndexOf);
     }
 }
