+++ src/main/java/org/snipsniper/utils/MathUtils.java
@@ -1,0 +1,21 @@
+package org.snipsniper.utils;
+
+public class MathUtils {
+    public static boolean isInteger(String string) {
+        try {
+            Integer.valueOf(string);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
+
+    public static boolean isDouble(String string) {
+        try {
+            Double.valueOf(string);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
+}
